% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{do_filt_out}
\alias{do_filt_out}
\title{Filter Out Url's}
\usage{
do_filt_out(url, regExOut)
}
\arguments{
\item{links}{Links to be filtered.}

\item{urlRegExFilterOut}{RegEX to filter with.}
}
\value{
filtered list of links
}
\description{
Filter out URL's
}
