% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_links}
\alias{extract_links}
\title{extract links}
\usage{
extract_links(doc, base_url)
}
\arguments{
\item{doc}{html document}

\item{base_url}{base url for extracted links}
}
\description{
takes html content and returns absolute url links
}
