% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_tags}
\alias{extract_tags}
\title{extract html tags}
\usage{
extract_tags(doc, tag)
}
\arguments{
\item{doc}{html document}

\item{tag}{tag to extract - h1,div,body,etc...}
}
\description{
takes html content and returns text from html tags
}
