% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_last_dir}
\alias{find_last_dir}
\title{Get Last Directory}
\usage{
find_last_dir(this = "fetch_", out_dir)
}
\arguments{
\item{this}{this directory}

\item{out_dir}{workind dir}
}
\value{
URLS beginning with (http/https), followed by ://www..
}
\description{
Get Last Directory
}
