% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_links}
\alias{get_links}
\title{Grab Links Found on Webpage.}
\usage{
get_links(res)
}
\arguments{
\item{res}{Return value from curl.}
}
\value{
Returns character vector of links.
}
\description{
Parses HTML for URLs and stores then in the linkDB.
}
