% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_optimal_res.R
\name{get_optimal_res}
\alias{get_optimal_res}
\title{Get optimal land mask raster resolution}
\usage{
get_optimal_res(pts)
}
\arguments{
\item{pts}{a SpatialPoints, or SpatialPointsDataFrame}
}
\value{
numeric value indicating the square cell dimension in meters
}
\description{
the \code{crawl} package includes a function \code{fix_path} that will
adjust a movement path to move around a restricted area (e.g. land if
the animal is a fish or marine mammal). The resolution of this land
mask is an important parameter and, often, is related to the
spatial scale of the movement. Higher resolution has a significant cost
in compute time and memory requirements. So, this function uses a simple
rule to determine an optimal resolution for this land mask. The highest
resolution is 200 m per side for any region with a max dimension less
than or equal to 25 km. Above a max dimension of 25 km, the resolution
is determined by dividing the max dimension by 1000 and multiplying by 8.
That said, once the max dimension reaches 150km, the resolution is set to
a fixed value of 1.2km.
}

