% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_root_from_url}
\alias{get_root_from_url}
\title{Returns the Root of a Url}
\usage{
get_root_from_url(urls)
}
\arguments{
\item{urls}{List of urls.}
}
\value{
Returns filename of output.
}
\description{
Internal use by crawlR package.
}
