% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim_tracks.R
\name{get_sim_tracks}
\alias{get_sim_tracks}
\title{Get All the Simulated Tracks You Want}
\usage{
get_sim_tracks(crw_obj, iter)
}
\arguments{
\item{crw_obj}{a crwFit object}

\item{iter}{the number of simulated tracks to return}
}
\value{
a list of length iter containing crwIS objects
}
\description{
Get All the Simulated Tracks You Want
}

