% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_url_from_root1}
\alias{get_url_from_root1}
\title{Returns New URL}
\usage{
get_url_from_root1(new_root, links, wait_table)
}
\arguments{
\item{new_root}{Root to draw next url from.}
}
\value{
Returns filename of output.
}
\description{
Internal use by crawlR package.
}
