% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injectR.R
\name{injectR}
\alias{injectR}
\title{Inject seeds into crawlDB}
\usage{
injectR(out_dir = NULL, work_dir = NULL, seeds = NULL, log_file = NULL)
}
\arguments{
\item{out_dir}{(Required) Output Directory directory for this crawl.}

\item{work_dir}{(Required) Working Directory directory for this crawl.}

\item{seeds}{(Required) Url's to inject.}

\item{log_file}{Name of log file. If null, writes to stdout().}
}
\value{
Returns character vector of links, or error message.
}
\description{
Inject seeds into crawlDB. If no crawlDB is found in \emph{work_dir}, a crawlDB will be created.

Can be called alone or thru the crawlR function. Seeds can be injected at anytime.
}
