% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_batch}
\alias{load_batch}
\title{Queue a Batch of URL's}
\usage{
load_batch(batch, queue, fetch_list)
}
\arguments{
\item{batch}{Batch Number to Load. Should start at 1.}

\item{queue}{Batch is loaeded into queue.}

\item{fetch_list}{Stores url's and their batch number.}
}
\value{
fetchable list
}
\description{
Queue a Batch of URL's
}
