% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{makeHash}
\alias{makeHash}
\title{Hashes URL to Make Valide (and unique) File Names}
\usage{
makeHash(url)
}
\arguments{
\item{url}{Url of crawled page.}

\item{type}{content-type of crawled page.}
}
\value{
Returns character vector of links, or error message.
}
\description{
Used by file output.
}
