% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_contact}
\alias{parse_contact}
\title{Parse Contact}
\usage{
parse_contact(res)
}
\arguments{
\item{res}{Return value from curl.}
}
\value{
Returns character vector of links, or error message.
}
\description{
Taken from crawl example given in the curl package.
}
