% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_content_fetch}
\alias{parse_content_fetch}
\title{General Parser}
\usage{
parse_content_fetch(res, readability, readability_content = F, map_meta = NULL)
}
\arguments{
\item{res}{Return value from curl.}

\item{readability}{readability reticulate object.}

\item{readability_content}{T of F - if true use readability.}

\item{map_meta}{meta data header mappings.}
}
\value{
Returns character vector of links, or error message.
}
\description{
Extracts the title, headers, span, and p tags from a page.
}
