% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parseExt}
\alias{parseExt}
\title{Get file extension from Content-Type}
\usage{
parseExt(type, extLookUp)
}
\arguments{
\item{type}{content-type of crawled page.}

\item{extLookUp}{file extension lookup.}
}
\value{
extention as string
}
\description{
Uses lookup provided to get file extension.
content-type: text/html would return ".html" file extension, and
content-type: application/pdf would return ".pdf".
}
