% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parseR}
\alias{parseR}
\title{Parse Processor}
\usage{
parseR(this_dir = NULL, parser = parse_content, log_file = NULL)
}
\arguments{
\item{this_dir}{Fetch directory to parse.}

\item{parser}{Function that will be used for parseing.}

\item{log_file}{Log file name. If null, defaults to stdout().}
}
\value{
URLS beginning with (http/https), followed by ://www..
}
\description{
Handles parseing of extracted links and page content.  Page content
is parsed using 'parser' function passed in as variable.
}
