% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parseR_2}
\alias{parseR_2}
\title{Parse Processor}
\usage{
parseR_2(this_dir = NULL, parser = parse_content, log_file = NULL)
}
\arguments{
\item{this_dir}{Fetch directory to parse.}

\item{log_file}{Log file name. If null, defaults to stdout().}

\item{parseR}{Function that will be used for parseing.}

\item{n_threads}{Number of threads to use for parseing.}

\item{n}{chunk size}
}
\value{
URLS beginning with (http/https), followed by ://www..
}
\description{
Parse Processor
}
