% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parser_wrapper}
\alias{parser_wrapper}
\title{Handles extracting links and applying supplied parse function.}
\usage{
parser_wrapper(
  res = NULL,
  parser = crawlR::parse_content,
  readability,
  readability_content = F
)
}
\arguments{
\item{res}{Return value from curl.}

\item{parser}{Supplied parser function.}

\item{readability}{not used}

\item{readability_content}{not used}
}
\value{

}
\description{
Extracts the links and passes the html doc to the given parser function.
}
