% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{process_fetched_links}
\alias{process_fetched_links}
\title{Processes Fetched Link List}
\usage{
process_fetched_links(fetched_links, fetch_list)
}
\arguments{
\item{fetched_links}{Links from fetch phase.}
}
\value{
fetchable list
}
\description{
Processes Fetched Link List
}
