% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{score_urls}
\alias{score_urls}
\title{Score urls}
\usage{
score_urls(
  paths = NULL,
  terms = c("about", "news", "press", "portfolio", "products", "announce", "construct",
    "plant", "coal", "nuclear", "wind", "refinery", "pipelines", "oil", "gas", "energy",
    "manufacturing", "chemical", "processing", "mining", "ore", "drilling", "project",
    "addition", "expansion", "grassroot", "greenfield"),
  scores = c(rep(1, 6), rep(1, 16), rep(1, 5))
)
}
\arguments{
\item{paths}{url paths to score}

\item{terms}{terms to use in scoring}

\item{scores}{values associated with terms}
}
\description{
Sets the out file for logs.
}
