% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-crwIS-methods.R
\docType{methods}
\name{subset-crwIS-methods}
\alias{subset-crwIS-methods}
\title{Subset \code{"crwIS"} objects}
\usage{
\S4method{subset}{crwIS}(x, locType = c("o", "p"), ...)
}
\description{
Package \pkg{crawlr} implements \code{subset}-methods to
subset \code{"crwIS"} S3 objects from
the \code{crawl} package based on \code{locType}
}
\author{
Josh M. London
}

