% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tika_mimetype}
\alias{tika_mimetype}
\title{Mimetypes from tika}
\format{
An object of class \code{data.frame} with 853 rows and 4 columns.
}
\usage{
tika_mimetype
}
\description{
Mimetypes from tika
}
\keyword{datasets}
