% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_iso8601.R
\name{to_iso8601}
\alias{to_iso8601}
\title{Convert POSIXct to an ISO 8601 string}
\usage{
to_iso8601(d, col, ...)
}
\arguments{
\item{d}{data frame containing a POSIXct column to convert}

\item{col}{POSIXct column to convert}

\item{...}{}
}
\value{
a data frame with the previous POSIXct column now an ISO 8601 string
}
\description{
Convert POSIXct to an ISO 8601 string
}

