% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateR.R
\name{updateR}
\alias{updateR}
\title{Update linkDB and pgDB After Crawl}
\usage{
updateR(
  work_dir = NULL,
  out_dir = NULL,
  crawl_delay = 5,
  log_file = NULL,
  overwrite = F,
  score_func = NULL
)
}
\arguments{
\item{work_dir}{(Required) Working directory for this crawl.}

\item{out_dir}{(Required) Working directory for this crawl.}

\item{crawl_delay}{crawl_delay}

\item{log_file}{Name of log file. If null, writes to stdout().}

\item{overwrite}{If true, all crawled links will be entered into DB as fresh unseen links.}

\item{score_func}{Function to score urls}

\item{max_concurr}{max conn.}
}
\value{
Returns character vector of links, or error message.
}
\description{
Updates the various DB's by taking the last fetched_pg file.
}
