% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{write_log}
\alias{write_log}
\title{Write to log}
\usage{
write_log(message, log_file)
}
\arguments{
\item{message}{If null, defaults to stdout().}

\item{log_file}{If null, defaults to stdout().}
}
\description{
Writes to log file.
}
