% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{writeR}
\alias{writeR}
\title{Base Output Writer (depricated)}
\usage{
writeR(out_dir = NULL, res = NULL, compress = T, extLookUp = NULL)
}
\arguments{
\item{out_dir}{Directory where output is written.}

\item{res}{Crawled page data.}

\item{compress}{If true, text/html is compressed.}

\item{extLookUp}{lookup file extension}
}
\value{
Returns filename of output.
}
\description{
Write web page to disk. Return file name of output.
}
