% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_fetch_list}
\alias{create_fetch_list}
\title{Creates a Fetch List from linkDB File}
\usage{
create_fetch_list(fetch_list, crawl_delay)
}
\arguments{
\item{linkDB}{linkDB}
}
\value{
fetchable list
}
\description{
Creates a Fetch List from linkDB File
}
