% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim_points.R
\name{get_sim_points}
\alias{get_sim_points}
\title{SpatialPoints from a list of Simulated Tracks}
\usage{
get_sim_points(sim_tracks, locType = c("p", "o"))
}
\arguments{
\item{locType}{}
}
\value{
a SpatialPointsDataFrame containing all points from sim_tracks
}
\description{
SpatialPoints from a list of Simulated Tracks
}

