% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{makeHash}
\alias{makeHash}
\title{Convert String to hash}
\usage{
makeHash(url)
}
\arguments{
\item{url}{Url of crawled page.}
}
\value{
Returns character vector of links, or error message.
}
\description{
uses openssl::md5 to convert url's to md5-hashes.
}
