% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{normalize_url}
\alias{normalize_url}
\title{Normalize Url's}
\usage{
normalize_url(url)
}
\arguments{
\item{url}{url to normalize}
}
\value{
URLS beginning with (http/https), followed by ://www..
}
\description{
Converts url's to lowercase, removes any trailing back slashes,
appends 'http://' to front if missing, and removes everything
following #-marks.
}
