% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_content}
\alias{parse_content}
\title{General Parser}
\usage{
parse_content(res)
}
\arguments{
\item{res}{Return value from curl.}
}
\value{
Returns character vector of links, or error message.
}
\description{
Extracts the title, headers, span, and p tags from a page.
}
