% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_content}
\alias{parse_content}
\title{General Parser for HTML}
\usage{
parse_content(doc)
}
\arguments{
\item{doc}{html content returned by xml2.}
}
\value{

}
\description{
Extracts the title, h1 headers, major meta-tags, and body tags from a page.
}
