% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parseR}
\alias{parseR}
\title{Parse Processor}
\usage{
parseR(
  this_dir = NULL,
  parser = parse_content,
  n_threads = 4,
  log_file = NULL,
  n = 4000
)
}
\arguments{
\item{this_dir}{Fetch directory to parse.}

\item{n_threads}{Number of threads to use for parseing.}

\item{log_file}{Log file name. If null, defaults to stdout().}

\item{n}{chunk size}

\item{parseR}{Function that will be used for parseing.}
}
\value{
URLS beginning with (http/https), followed by ://www..
}
\description{
Parse Processor
}
