% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{score_urls}
\alias{score_urls}
\title{Score urls}
\usage{
score_urls(
  paths = NULL,
  terms = c(),
  scores = c(rep(1, 6), rep(1, 16), rep(1, 5))
)
}
\arguments{
\item{paths}{url paths to score}

\item{terms}{terms to use in scoring}

\item{scores}{values associated with terms}
}
\description{
Sets the out file for logs.
}
