% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ud_hexpolys.R
\name{ud_hexpolys}
\alias{ud_hexpolys}
\title{Create Hexagonal Utilization Distribution Polygons}
\usage{
ud_hexpolys(data_sp, cellsize, leaflet = FALSE, density = TRUE,
  seed = 123)
}
\arguments{
\item{data_sp}{SpatialPointsDataFrame of regular locations}

\item{cellsize}{cellsize parameter passed to spsample}

\item{leaflet}{should the output be converted to EPSG:4326 for leaflet}

\item{seed}{for consistency}
}
\value{
a SpatialPolygonsDataFrame
}
\description{
Create Hexagonal Utilization Distribution Polygons
}
\examples{
# library(crawl)
}

